+++ src/Groups/GroupsList/GroupsList.js
@@ -8,7 +8,10 @@ import {
 } from 'react-router-dom';
 import { FormattedMessage } from 'react-intl';
 
-import { useStripes } from '@folio/stripes/core';
+import {
+  TitleManager,
+  useStripes,
+} from '@folio/stripes/core';
 import {
   checkScope,
   HasCommand,
@@ -32,7 +35,10 @@ import {
 } from '@folio/stripes-acq-components';
 
 import { GROUPS_ROUTE } from '../../common/const';
-import { useSelectedRow } from '../../common/hooks';
+import {
+  useResultsPageTitle,
+  useSelectedRow,
+} from '../../common/hooks';
 import FinanceNavigation from '../../common/FinanceNavigation';
 import CheckPermission from '../../common/CheckPermission';
 
@@ -84,6 +90,7 @@ const GroupsList = ({
     changeSorting,
   ] = useLocationSorting(location, history, resetData, sortableFields);
 
+  const pageTitle = useResultsPageTitle(filters);
   const { isFiltersOpened, toggleFilters } = useFiltersToogle('ui-finance/group/filters');
   const { itemToView, setItemToView, deleteItemToView } = useItemToView('groups-list');
 
@@ -122,6 +129,7 @@ const GroupsList = ({
       isWithinScope={checkScope}
       scope={document.body}
     >
+      <TitleManager page={pageTitle} />
       <PersistedPaneset
         appId="ui-finance"
         id="group-paneset"