+++ src/Ledger/LedgerList/LedgersList.js
@@ -9,7 +9,10 @@ import {
   FormattedMessage,
 } from 'react-intl';
 
-import { useStripes } from '@folio/stripes/core';
+import {
+  TitleManager,
+  useStripes,
+} from '@folio/stripes/core';
 import {
   checkScope,
   HasCommand,
@@ -33,7 +36,10 @@ import {
 } from '@folio/stripes-acq-components';
 
 import { LEDGERS_ROUTE } from '../../common/const';
-import { useSelectedRow } from '../../common/hooks';
+import {
+  useResultsPageTitle,
+  useSelectedRow,
+} from '../../common/hooks';
 import FinanceNavigation from '../../common/FinanceNavigation';
 import CheckPermission from '../../common/CheckPermission';
 import LedgerListFilters from './LedgerListFilters';
@@ -85,6 +91,7 @@ const LedgerList = ({
 
   useFiltersReset(resetFilters);
 
+  const pageTitle = useResultsPageTitle(filters);
   const { isFiltersOpened, toggleFilters } = useFiltersToogle('ui-finance/ledger/filters');
 
   const renderLastMenu = useCallback(() => <LedgerListLastMenu />, []);
@@ -125,6 +132,7 @@ const LedgerList = ({
       isWithinScope={checkScope}
       scope={document.body}
     >
+      <TitleManager page={pageTitle} />
       <PersistedPaneset
         appId="ui-receiving"
         id="invoice-paneset"