+++ src/common/hooks/index.js
@@ -3,4 +3,5 @@ export { useFundTypesBatch } from './useFundTypesBatch';
 export { useLedger } from './useLedger';
 export { useLedgerRollover } from './useLedgerRollover';
 export { useOrder } from './useOrder';
+export { useResultsPageTitle } from './useResultsPageTitle';
 export { useSelectedRow } from './useSelectedRow';