+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Incorrect Display of Expense Class Percent. Refs UIF-509.
 * Make location association required when funds are restricted by location. Refs UIF-492.
 * *BREAKING* Add "Unrelease encumbrance" action to encumbrance transaction view. Refs UIF-503.
+* UX Consistency > Search results > Update HTML page title with search term entered. Refs UIF-501.
 
 ## [5.0.0](https://github.com/folio-org/ui-finance/tree/v5.0.0) (2023-10-12)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v4.0.4...v5.0.0)